Extract song property functions
